Rename/Merge CascadingSettings to AppSettings
