--- src/Web.App/Settings/AppSettings.cs
@@ -1,11 +1,0 @@
-// Copyright (c) KappaDuck. All rights reserved.
-// The source code is licensed under MIT License.
-
-namespace Web.App.Settings;
-
-internal sealed record AppSettings
-{
-    public Theme Theme { get; init; }
-
-    public Language Language { get; init; }
-}
