+++ src/Web.App/Settings/AppSettings.razor
@@ -1,0 +1,3 @@
+﻿<CascadingValue Value="this" IsFixed="true">
+    @ChildContent
+</CascadingValue>
