--- src/Web.App/Settings/CascadingSettings.razor
@@ -1,3 +1,0 @@
-﻿<CascadingValue Value="_settings" IsFixed="false">
-    @ChildContent
-</CascadingValue>
